Moved some stuff to the annotation class
