+++ Symfony/CS/DocBlock/Annotation.php
@@ -19,13 +19,35 @@ namespace Symfony\CS\DocBlock;
 class Annotation
 {
     /**
-     * The lines that make up the annotation.
+     * All the annotation tag names with types.
      *
-     * Note that the array indexes represent the position in the docblock.
+     * @var string[]
+     */
+    private static $tags = array(
+        'method', 'param', 'property', 'property-read', 'property-write',
+        'return', 'throws', 'type', 'var',
+    );
+
+    /**
+     * The lines that make up the annotation.
      *
      * @var Line[]
      */
     private $lines;
+
+    /**
+     * The position of the first line of the annotation in the docblock.
+     *
+     * @var int
+     */
+    private $start;
+
+    /**
+     * The position of the last line of the annotation in the docblock.
+     *
+     * @var int
+     */
+    private $end;
 
     /**
      * The associated tag.
@@ -35,13 +57,25 @@ class Annotation
     private $tag;
 
     /**
+     * The cached types content.
+     *
+     * @var string|null
+     */
+    private $typesContent;
+
+    /**
      * Create a new line instance.
      *
      * @param Line[] $lines
      */
     public function __construct(array $lines)
     {
-        $this->lines = $lines;
+        $this->lines = array_values($lines);
+
+        $keys = array_keys($lines);
+
+        $this->start = $keys[0];
+        $this->end = end($keys);
     }
 
     /**
@@ -51,9 +85,7 @@ class Annotation
      */
     public function getStart()
     {
-        $keys = array_keys($this->lines);
-
-        return $keys[0];
+        return $this->start;
     }
 
     /**
@@ -63,9 +95,7 @@ class Annotation
      */
     public function getEnd()
     {
-        $keys = array_keys($this->lines);
-
-        return end($keys);
+        return $this->end;
     }
 
     /**
@@ -76,11 +106,59 @@ class Annotation
     public function getTag()
     {
         if (null === $this->tag) {
-            $values = array_values($this->lines);
-            $this->tag = new Tag($values[0]);
+            $this->tag = new Tag($this->lines[0]);
         }
 
         return $this->tag;
+    }
+
+    /**
+     * Get the current types content.
+     *
+     * Be careful modifying the underlying line as that won't flush the cache.
+     *
+     * @return string
+     */
+    private function getTypesContent()
+    {
+        if (null === $this->typesContent) {
+            $name = $this->getTag()->getName();
+
+            if (!in_array($name, self::$tags, true)) {
+                throw new \RuntimeException('This tag does not support types');
+            }
+
+            $tagSplit = preg_split('/\s*\@'.$name.'\s*/', $this->lines[0]->getContent(), 2);
+            $spaceSplit = preg_split('/\s/', $tagSplit[1], 2);
+
+            $this->typesContent = $spaceSplit[0];
+        }
+
+        return $this->typesContent;
+    }
+
+    /**
+     * Get the types associated with this annotation.
+     *
+     * @return string[]
+     */
+    public function getTypes()
+    {
+        return explode('|', $this->getTypesContent());
+    }
+
+    /**
+     * Set the types associated with this annotation.
+     *
+     * @param string[] $types
+     */
+    public function setTypes(array $types)
+    {
+        $pattern = '/'.preg_quote($this->getTypesContent()).'/';
+
+        $this->lines[0]->setContent(preg_replace($pattern, implode('|', $types), $this->lines[0]->getContent(), 1));
+
+        $this->typesContent = null;
     }
 
     /**
