+++ Symfony/CS/DocBlock/Tag.php
@@ -32,16 +32,6 @@ class Tag
     );
 
     /**
-     * All the tags with types.
-     *
-     * @var string[]
-     */
-    private static $tagsWithTypes = array(
-        'method', 'param', 'property', 'property-read', 'property-write',
-        'return', 'throws', 'type', 'var',
-    );
-
-    /**
      * The line containing the tag.
      *
      * @var Line
@@ -54,13 +44,6 @@ class Tag
      * @var string|null
      */
     private $name;
-
-    /**
-     * The cached types content.
-     *
-     * @var string|null
-     */
-    private $typesContent;
 
     /**
      * Create a new tag instance.
@@ -124,54 +107,5 @@ class Tag
         $this->line->setContent(preg_replace("/@$current/", "@$name", $this->line->getContent(), 1));
 
         $this->name = $name;
-    }
-
-    /**
-     * Get the types associated with this tag.
-     *
-     * @return string[]
-     */
-    public function getTypes()
-    {
-        return explode('|', $this->getTypesContent());
-    }
-
-    /**
-     * Set the types associated with this tag.
-     *
-     * @param string[] $types
-     */
-    public function setTypes(array $types)
-    {
-        $pattern = '/'.preg_quote($this->getTypesContent()).'/';
-
-        $this->line->setContent(preg_replace($pattern, implode('|', $types), $this->line->getContent(), 1));
-
-        $this->typesContent = null;
-    }
-
-    /**
-     * Get the current types content.
-     *
-     * Be careful modifying the underlying line as that won't flush the cache.
-     *
-     * @return string
-     */
-    private function getTypesContent()
-    {
-        if (null === $this->typesContent) {
-            $name = $this->getName();
-
-            if (!in_array($this->getName(), self::$tagsWithTypes, true)) {
-                throw new \RuntimeException('This tag does not support types');
-            }
-
-            $tagSplit = preg_split('/\s*\@'.$name.'\s*/', $this->line->getContent(), 2);
-            $spaceSplit = preg_split('/\s/', $tagSplit[1], 2);
-
-            $this->typesContent = $spaceSplit[0];
-        }
-
-        return $this->typesContent;
     }
 }
