+++ Symfony/CS/Tests/DocBlock/TagTest.php
@@ -79,53 +79,4 @@ class TagTest extends \PHPUnit_Framework_TestCase
             array('other', 'hello', ' * @method("GET")'),
         );
     }
-
-    /**
-     * @dataProvider provideTypesCases
-     */
-    public function testTypes($expected, $new, $input, $output)
-    {
-        $line = new Line($input);
-        $tag = new Tag($line);
-
-        $this->assertSame($expected, $tag->getTypes());
-
-        $tag->setTypes($new);
-
-        $this->assertSame($new, $tag->getTypes());
-
-        $this->assertSame($output, $line->getContent());
-    }
-
-    public function provideTypesCases()
-    {
-        return array(
-            array(array('Foo', 'null'), array('Bar[]'), '     * @param Foo|null $foo', '     * @param Bar[] $foo'),
-            array(array('false'), array('bool'), '*   @return            false', '*   @return            bool'),
-            array(array('RUNTIMEEEEeXCEPTION'), array('Throwable'), "\t@throws\t  \t RUNTIMEEEEeXCEPTION\t\t\t\t\t\t\t\n\n\n", "\t@throws\t  \t Throwable\t\t\t\t\t\t\t\n\n\n"),
-            array(array('string'), array('string', 'null'), ' * @method string getString()', ' * @method string|null getString()'),
-        );
-    }
-
-    /**
-     * @expectedException RuntimeException
-     * @expectedExceptionMessage This tag does not support types
-     */
-    public function testGetTypesOnBadTag()
-    {
-        $tag = new Tag(new Line(' * @deprecated since 1.2'));
-
-        $tag->getTypes();
-    }
-
-    /**
-     * @expectedException RuntimeException
-     * @expectedExceptionMessage This tag does not support types
-     */
-    public function testSetTypesOnBadTag()
-    {
-        $tag = new Tag(new Line(' * @author Chuck Norris'));
-
-        $tag->setTypes(array('string'));
-    }
 }
